requirement.txt
asgiref==3.2.10
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
gunicorn==20.0.4
idna==2.9
mysqlclient==1.4.6
numpy==1.18.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
whitenoise==5.1.0
